New df, and more.
Added a new dataframe which will hold the coronavirus data that we can visually compare our original dataframe to. Additionally, added annotations for January 14th, 2020 after convering the values to datetime for "Case #1" according to the CDC.

Added a completely new plot that compares all social media platforms on the same plot in relation to coronavirus on it's own plot.